refactor(renderer): extract GTAO compute into hiz.rs (record_gtao)
Cluster 9 prep: the GTAO dispatch (temporal EMA ping-pong, Halton-5
rotation) leaves end_frame_with_scene; p22/p32 derive inside the
method. Goldens pixel-identical.
